pyproject.toml
[tool.poetry]
name = "alpaca-calls-ws"
version = "0.1.0"
description = ""
authors = ["cashblxxd <cashblxxd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.3.3"
alpaca-py = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
